Removes v0 location transformers  (#25025)
* Removes v0 location transformer functions


* Removes v0 location transformer unit tests


* Added functions still used to the v2 location transformers


* Commenting this out, the entire file is removed in another open PR, allows tests to pass


* Remames v2 transformers file


* Temporarily fixes import


* Revert "Temporarily fixes import"

This reverts commit 1d9a8dab08b6d42d36d2a72b202455617f9e210f.


---------
